feat: created a error handler for the api.
