+++ firstjavacrud/src/main/java/com/restapi/firstjavacrud/exceptions/ApiExceptionHandler.java
@@ -1,23 +1,34 @@
 package com.restapi.firstjavacrud.exceptions;
 
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
+/**
+ * When an ApiRequestException is thrown, return a ResponseEntity with the
+ * exception's message and
+ * timestamp, and the exception's HTTP status.
+ */
 @ControllerAdvice
+/**
+ * This class is a controller advice that handles exceptions thrown by the
+ * controller
+ */
 public class ApiExceptionHandler {
-    private HttpStatus badRequest = HttpStatus.BAD_REQUEST;
-
+    /**
+     * This function is handling the exception thrown by the controller
+     * 
+     * @param exception The exception object that was thrown by the controller.
+     * @return The response entity is being returned.
+     */
     @ExceptionHandler(value = { ApiRequestException.class })
-    public ResponseEntity<Object> handleApiExceptionHandler(ApiRequestException e) {
-        ApiException apiException = new ApiException(e.getLocalizedMessage(),
-                // e,
-                badRequest,
-                ZonedDateTime.now(ZoneId.of("America/Sao_Paulo")));
-        return new ResponseEntity<>(apiException, badRequest);
+    // This method is handling the exception thrown by the controller.
+    public ResponseEntity<Object> handleApiException(Exception exception) {
+        ApiRequestException exceptionArgs = (ApiRequestException) exception;
+        return new ResponseEntity<>(
+                new ApiExceptionResponse(
+                        exceptionArgs.getMessage(),
+                        exceptionArgs.getTimestamp()),
+                exceptionArgs.getHttpStatus());
     }
 }
